+++ src/components/dlc/platforms/AmagProTime/AmagProTime.content.js
@@ -2,11 +2,22 @@ export const customContent = `<div class="dlc-platform_custom-content">
     <p class="text-label">Funktionen</p>
     <div class="dlc-platform_custom-content-row dFlex">
         <div class="row-cell-left">
-            <p class="subtext">ProTime Test </p>
+            <p class="subtext">ProTime Test</p>
         </div>
         <div class="row-cell-right">
             <label class="switch">
                 <input type="checkbox" class="" id="check_showProTimetestButton" />
+                <span class="switch_slider"></span>
+            </label>
+        </div>
+    </div>
+    <div class="dlc-platform_custom-content-row dFlex">
+        <div class="row-cell-left">
+            <p class="subtext">Force Latency-Mode</p>
+        </div>
+        <div class="row-cell-right">
+            <label class="switch">
+                <input type="checkbox" class="" id="check_forceLatencyModeproTime" />
                 <span class="switch_slider"></span>
             </label>
         </div>